refactor/ rename PREFIX -> POSTFIX
